+++ sfml-util/gui/creature-image-manager.cpp
@@ -105,7 +105,7 @@ namespace gui
 
                             M_ASSERT_OR_LOGANDTHROW_SS((filenamesVec.empty() == false), "sfml_util::gui::CreatureImageManager() (wolfen_classes) race=" << RACE_STR << ", role=" << ROLE_STR << ", sex=" << SEX_STR << ", wolfen_class=" << CLASS_STR << ", GetFilenames() failed to return anything.");
 
-                            if (i < static_cast<int>(filenamesVec.size()))
+                            if (i < filenamesVec.size())
                             {
                                 auto const TEXTURE_SPTR{ cimSPtr->GetImage(filenamesVec.at(i)) };
                                 M_ASSERT_OR_LOGANDTHROW_SS((TEXTURE_SPTR.get() != nullptr), "sfml_util::gui::CreatureImageManager() (wolfen_classes) race=" << RACE_STR << ", role=" << ROLE_STR << ", sex=" << SEX_STR << ", wolfen_class=" << CLASS_STR << ", GetImage(\"" << filenamesVec.at(i) << "\") returned a nullptr texture.");
@@ -139,7 +139,7 @@ namespace gui
 
                                 M_ASSERT_OR_LOGANDTHROW_SS((filenamesVec.empty() == false), "sfml_util::gui::CreatureImageManager() (dragon_classes) race=" << RACE_STR << ", role=" << ROLE_STR << ", sex=" << SEX_STR << ", dragon_class=" << CLASS_STR << ", GetFilenames() failed to return anything.");
 
-                                if (i < static_cast<int>(filenamesVec.size()))
+                                if (i < filenamesVec.size())
                                 {
                                     auto const TEXTURE_SPTR{ cimSPtr->GetImage(filenamesVec.at(i)) };
                                     M_ASSERT_OR_LOGANDTHROW_SS((TEXTURE_SPTR.get() != nullptr), "sfml_util::gui::CreatureImageManager() (dragon_classes) race=" << RACE_STR << ", role=" << ROLE_STR << ", sex=" << SEX_STR << ", dragon_class=" << CLASS_STR << ", GetImage(\"" << filenamesVec.at(i) << "\") returned a nullptr texture.");
@@ -167,7 +167,7 @@ namespace gui
 
                         M_ASSERT_OR_LOGANDTHROW_SS((filenamesVec.empty() == false), "sfml_util::gui::CreatureImageManager() race=" << RACE_STR << ", role=" << ROLE_STR << ", sex=" << SEX_STR << ", GetFilenames() failed to return anything.");
 
-                        if (i < static_cast<int>(filenamesVec.size()))
+                        if (i < filenamesVec.size())
                         {
                             auto const TEXTURE_SPTR{ cimSPtr->GetImage(filenamesVec.at(i)) };
                             M_ASSERT_OR_LOGANDTHROW_SS((TEXTURE_SPTR.get() != nullptr), "sfml_util::gui::CreatureImageManager() (dragon_classes) race=" << RACE_STR << ", role=" << ROLE_STR << ", sex=" << SEX_STR << ", GetImage(\"" << filenamesVec.at(i) << "\") returned a nullptr texture.");
@@ -258,7 +258,7 @@ namespace gui
 
         if (WILL_PICK_RANDOM && (filenameVec.size() > 1))
         {
-            return filenameVec[static_cast<std::size_t>(utilz::random::Int(0, filenameVec.size() - static_cast<std::size_t>(1)))];
+            return filenameVec[static_cast<std::size_t>(utilz::random::Int(0, static_cast<int>(filenameVec.size()) - 1))];
         }
         else
         {
